backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.3.0
build==0.10.0
CacheControl==0.12.14
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.0.1
cleo==2.0.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
crashtest==0.4.1
cryptography==41.0.4
daphne==4.0.0
distlib==0.3.6
Django==4.2.6
django-cors-headers==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
dulwich==0.21.5
filelock==3.12.2
html5lib==1.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.6.0
incremental==22.10.0
installer==0.7.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pexpect==4.8.0
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.5.3
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pytz==2023.3.post1
rapidfuzz==2.15.1
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
service-identity==23.1.0
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
tbomb==2.1.2
tomlkit==0.11.8
trove-classifiers==2023.5.24
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==1.26.16
virtualenv==20.23.0
webencodings==0.5.1
xattr==0.10.1
yarg==0.1.9
zipp==3.15.0
zope.interface==6.1
